+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Plume-Util change log
 
-## 2.0.1 (2026-08-??)
+## 2.1.0 (2026-??-??)
 
 - Requires Java 21 to run.
 - Renamed final static fields to SCREAMING_CASE, such as `it`.
   The old names remain but are deprecated.
+- Converted classes to records; their fields are now accessor methods:
+  - `EntryReader.EntryFormat`
+  - `EntryReader.CommentFormat`
 
 ## 2.0.0 (2026-08-05)